Use shared period helpers in habit queries
