+++ src/app/actions/habits/update.ts
@@ -2,23 +2,11 @@
 
 import { Result } from '@praha/byethrow'
 import { revalidatePath } from 'next/cache'
-import { NotFoundError, UnauthorizedError } from '@/lib/errors/habit'
+import { NotFoundError } from '@/lib/errors/habit'
 import { serializeHabitError } from '@/lib/errors/serializable'
 import { updateHabit } from '@/lib/queries/habit'
-import { getCurrentUserId } from '@/lib/user'
 import { validateHabitUpdate } from '@/validators/habit'
-
-/**
- * 認証チェック
- * @returns Result<userId, UnauthorizedError>
- */
-const authenticateUser = async (): Result.ResultAsync<string, UnauthorizedError> => {
-  const userId = await getCurrentUserId()
-  if (!userId) {
-    return Result.fail(new UnauthorizedError())
-  }
-  return Result.succeed(userId)
-}
+import { authenticateUser } from './action-utils'
 
 /**
  * 習慣を更新するServer Action
