+++ src/lib/queries/period.ts
@@ -1,0 +1,53 @@
+import { endOfDay, endOfMonth, endOfWeek, startOfDay, startOfMonth, startOfWeek } from 'date-fns'
+import type { Period, WeekStartDay } from '@/constants/habit'
+import { formatDateKey, parseDateKey } from '@/lib/utils/date'
+
+/**
+ * 期間の開始日時を計算
+ */
+export const getPeriodStart = (date: Date, period: Period, weekStartDay: WeekStartDay = 1): Date => {
+  switch (period) {
+    case 'daily':
+      return startOfDay(date)
+    case 'weekly': {
+      return startOfWeek(date, { weekStartsOn: weekStartDay })
+    }
+    case 'monthly':
+      return startOfMonth(date)
+    default:
+      return startOfDay(date)
+  }
+}
+
+/**
+ * 期間の終了日時を計算
+ */
+export const getPeriodEnd = (date: Date, period: Period, weekStartDay: WeekStartDay = 1): Date => {
+  switch (period) {
+    case 'daily':
+      return endOfDay(date)
+    case 'weekly': {
+      return endOfWeek(date, { weekStartsOn: weekStartDay })
+    }
+    case 'monthly': {
+      return endOfMonth(date)
+    }
+    default:
+      return endOfDay(date)
+  }
+}
+
+export const getPeriodBounds = (
+  date: Date | string,
+  period: Period,
+  weekStartDay: WeekStartDay = 1
+): { startKey: string; endKey: string } => {
+  const baseDate = typeof date === 'string' ? parseDateKey(date) : date
+  const start = getPeriodStart(baseDate, period, weekStartDay)
+  const end = getPeriodEnd(baseDate, period, weekStartDay)
+
+  return {
+    startKey: formatDateKey(start),
+    endKey: formatDateKey(end),
+  }
+}
