+++ src/types/user.ts
@@ -1,0 +1,7 @@
+export interface User {
+  id: string
+  clerkId: string
+  email: string
+  createdAt: Date
+  updatedAt: Date
+}
